+++ model/database.js
@@ -2,7 +2,7 @@ const { Sequelize} = require('sequelize');
 
 const sequelize = new Sequelize({
     dialect:'sqlite',
-    storage:'../database.splite'
+    storage:'./database.splite'
 });
   
 module.exports = sequelize;
